third_party/blink/web_tests/fast/dom/non-numeric-values-numeric-parameters.html

<!DOCTYPE HTML PUBLIC>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
</body>
<script>
"use strict";

// This tests the behavior of non-numeric values in contexts where the DOM has a
// numeric parameter.

async function nonNumericPolicy(template) {
  let x = 0;
  try {
    await eval(template);
  } catch (e) {
    return e;
  }

  let nullAllowed = 1;
  x = null;
  try {
    await eval(template);
  } catch (e) {
    nullAllowed = 0;
  }

  let undefinedAllowed = 1;
  x = undefined;
  try {
    await eval(template);
  } catch (e) {
    undefinedAllowed = 0;
  }

  let stringAllowed = 1;
  x = "string";
  try {
    await eval(template);
  } catch (e) {
    stringAllowed = 0;
  }

  let documentAllowed = 1;
  x = document;
  try {
    await eval(template);
  } catch (e) {
    documentAllowed = 0;
  }

  let nonIntegerAllowed = 1;
  x = 0.1;
  try {
    await eval(template);
  } catch (e) {
    nonIntegerAllowed = 0;
  }

  let infinityAllowed = 1;
  x = Infinity;
  try {
    await eval(template);
  } catch (e) {
    infinityAllowed = 0;
  }

  let nanAllowed = 1;
  x = NaN;
  try {
    await eval(template);
  } catch (e) {
    nanAllowed = 0;
  }

  let omitAllowed = -1; // means "not applicable"
  const templateWithoutArg = template.replace("x, ", "").replace(", x)", ")")
      .replace("(x)", "()");
  if (templateWithoutArg != template) {
    omitAllowed = 1;
    try {
      await eval(templateWithoutArg);
    } catch(e) {
      omitAllowed = 0;
    }
  }

  const expectOmitAllowed = navigator.userAgent.match("Gecko/") != "Gecko/";

  if (nullAllowed && undefinedAllowed && stringAllowed && documentAllowed &&
      nonIntegerAllowed && infinityAllowed && nanAllowed) {
    if (omitAllowed == -1 || omitAllowed == (expectOmitAllowed ? 1 : 0))
      return "any type allowed";
    if (omitAllowed == 1)
      return "any type allowed (or omitted)";
    if (omitAllowed == 0)
      return "any type allowed (but not omitted)";
  }
  if (nullAllowed && !undefinedAllowed && !stringAllowed && !documentAllowed &&
      nonIntegerAllowed && !infinityAllowed && nanAllowed && omitAllowed == 1)
    return "number or null allowed (or omitted, but not infinite)";
  return "mixed";
}

const selector = "a";
const styleText = "font-size: smaller";
const ruleText = selector + " { " + styleText + " }";

const testElementContainer = document.createElement("div");
document.body.appendChild(testElementContainer);

function createFromMarkup(markup) {
  const range = document.createRange();
  const fragmentContainer = document.createElement("div");
  range.selectNodeContents(fragmentContainer);
  testElementContainer.appendChild(fragmentContainer);
  const fragment = range.createContextualFragment(markup);
  fragmentContainer.appendChild(fragment);
  return fragmentContainer.firstChild;
}

function createCSSStyleSheet() {
  return createFromMarkup("<style>" + ruleText + "</style>").sheet;
}

function createCSSRuleList() {
  return createCSSStyleSheet().cssRules;
}

function createCSSStyleDeclaration() {
  return createCSSRuleList().item(0).style;
}

function createCSSMediaRule() {
  const rule = createFromMarkup("<style>@media screen { a { text-weight: bold } }</style>")
      .sheet.cssRules.item(0);
  rule.insertRule(ruleText, 0);
  return rule;
}

function createMediaList() {
  return createCSSMediaRule().media;
}

function createHTMLSelectElement() {
  const select = document.createElement("select");
  select.options.add(document.createElement("option"));
  return select;
}

function createHTMLOptionsCollection() {
  return createHTMLSelectElement().options;
}

function createHTMLTableElement() {
  const table = document.createElement("table");
  table.insertRow(0);
  return table;
}

function createHTMLTableSectionElement() {
  const table = document.createElement("table");
  table.insertRow(0);
  return table.tBodies[0];
}

function createHTMLTableRowElement() {
  const table = document.createElement("table");
  const row = table.insertRow(0);
  row.insertCell(0);
  return row;
}

function createCanvasElement() {
  return document.createElement("canvas");
}

promise_test(async () => {
  // CharacterData
  assert_equals(await nonNumericPolicy('document.createTextNode(\"a\").substringData(x, 0)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('document.createTextNode(\"a\").substringData(0, x)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('document.createTextNode(\"a\").insertData(x, \"b\")'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('document.createTextNode(\"a\").deleteData(x, 0)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('document.createTextNode(\"a\").deleteData(0, x)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('document.createTextNode(\"a\").replaceData(x, 0, \"b\")'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('document.createTextNode(\"a\").replaceData(0, x, \"b\")'),
      'any type allowed (but not omitted)');
}, "CharacterData");

promise_test(async () => {
  // CSSMediaRule
  assert_equals(await nonNumericPolicy('createCSSMediaRule().insertRule(ruleText, x)'),
      'any type allowed');
  assert_equals(await nonNumericPolicy('createCSSMediaRule().deleteRule(x)'),
      'any type allowed (but not omitted)');

  // CSSRuleList
  assert_equals(await nonNumericPolicy('createCSSRuleList().item(x)'),
      'any type allowed (but not omitted)');

  // CSSStyleDeclaration
  assert_equals(await nonNumericPolicy('createCSSStyleDeclaration().item(x)'),
      'any type allowed (but not omitted)');

  // CSSStyleSheet
  assert_equals(await nonNumericPolicy('createCSSStyleSheet().insertRule(ruleText, x)'),
      'any type allowed');
  assert_equals(await nonNumericPolicy('createCSSStyleSheet().deleteRule(x)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('createCSSStyleSheet().addRule(selector, styleText, x)'),
      'any type allowed');
  assert_equals(await nonNumericPolicy('createCSSStyleSheet().removeRule(x)'),
      'any type allowed');
}, "CSS*");

promise_test(async () => {
  // Document
  assert_equals(await nonNumericPolicy('document.elementFromPoint(x, 0)'),
      'mixed');
  assert_equals(await nonNumericPolicy('document.elementFromPoint(0, x)'),
      'mixed');
}, "Document");

promise_test(async () => {
  // Element
  assert_equals(await nonNumericPolicy('document.body.scrollLeft = x'),
      'any type allowed');
  assert_equals(await nonNumericPolicy('document.body.scrollTop = x'),
      'any type allowed');
}, "Element");

promise_test(async () => {
  // HTMLCollection
  assert_equals(await nonNumericPolicy('document.images.item(x)'),
      'any type allowed (but not omitted)');

  // HTMLInputElement
  assert_equals(await nonNumericPolicy('document.createElement(\"input\").setSelectionRange(x, 0)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('document.createElement(\"input\").setSelectionRange(0, x)'),
      'any type allowed (but not omitted)');

  // HTMLOptionsCollection
  assert_equals(await nonNumericPolicy('createHTMLOptionsCollection().add(document.createElement(\"option\"), x)'),
      'any type allowed');
  assert_equals(await nonNumericPolicy('createHTMLOptionsCollection().remove(x)'),
      'any type allowed (but not omitted)');

  // HTMLSelectElement
  assert_equals(await nonNumericPolicy('createHTMLSelectElement().remove(x)'),
      'any type allowed');
  assert_equals(await nonNumericPolicy('createHTMLSelectElement().item(x)'),
      'any type allowed (but not omitted)');

  // HTMLTableElement
  assert_equals(await nonNumericPolicy('createHTMLTableElement().insertRow(x)'),
      'any type allowed');
  assert_equals(await nonNumericPolicy('createHTMLTableElement().deleteRow(x)'),
      'any type allowed (but not omitted)');

  // HTMLTableRowElement
  assert_equals(await nonNumericPolicy('createHTMLTableRowElement().insertCell(x)'),
      'any type allowed');
  assert_equals(await nonNumericPolicy('createHTMLTableRowElement().deleteCell(x)'),
      'any type allowed (but not omitted)');

  // HTMLTableSectionElement
  assert_equals(await nonNumericPolicy('createHTMLTableSectionElement().insertRow(x)'),
      'any type allowed');
  assert_equals(await nonNumericPolicy('createHTMLTableSectionElement().deleteRow(x)'),
      'any type allowed (but not omitted)');

  // HTMLInputElement
  assert_equals(await nonNumericPolicy('document.createElement(\"textarea\").setSelectionRange(x, 0)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('document.createElement(\"textarea\").setSelectionRange(0, x)'),
      'any type allowed (but not omitted)');

  // HTMLCanvasElement
  assert_equals(await nonNumericPolicy('createCanvasElement().getContext(x)'),
      'any type allowed (but not omitted)');
}, "HTML*");

promise_test(async () => {
  // KeyboardEvent
  assert_equals(await nonNumericPolicy('document.createEvent(\"KeyboardEvent\").initKeyboardEvent(\"a\", false, false, null, \"b\", x, false, false, false, false, false)'),
      'any type allowed');
}, "KeyboardEvent");

promise_test(async () => {
  // MediaList
  assert_equals(await nonNumericPolicy('createMediaList().item(x)'),
      'any type allowed (but not omitted)');
}, "MediaList");

promise_test(async () => {
  // MouseEvent
  assert_equals(await nonNumericPolicy('document.createEvent(\"MouseEvent\").initMouseEvent(\"a\", false, false, null, x, 0, 0, 0, 0, false, false, false, false, 0, null)'),
      'any type allowed');
  assert_equals(await nonNumericPolicy('document.createEvent(\"MouseEvent\").initMouseEvent(\"a\", false, false, null, 0, x, 0, 0, 0, false, false, false, false, 0, null)'),
      'any type allowed');
  assert_equals(await nonNumericPolicy('document.createEvent(\"MouseEvent\").initMouseEvent(\"a\", false, false, null, 0, 0, x, 0, 0, false, false, false, false, 0, null)'),
      'any type allowed');
  assert_equals(await nonNumericPolicy('document.createEvent(\"MouseEvent\").initMouseEvent(\"a\", false, false, null, 0, 0, 0, x, 0, false, false, false, false, 0, null)'),
      'any type allowed');
  assert_equals(await nonNumericPolicy('document.createEvent(\"MouseEvent\").initMouseEvent(\"a\", false, false, null, 0, 0, 0, 0, x, false, false, false, false, 0, null)'),
      'any type allowed');
  assert_equals(await nonNumericPolicy('document.createEvent(\"MouseEvent\").initMouseEvent(\"a\", false, false, null, 0, 0, 0, 0, 0, false, false, false, false, x, null)'),
      'any type allowed');
}, "MouseEvent");

promise_test(async () => {
  // NamedNodeMap
  assert_equals(await nonNumericPolicy('document.body.attributes.item(x)'),
      'any type allowed (but not omitted)');

  // NodeIterator
  assert_equals(await nonNumericPolicy('document.createNodeIterator(document, x, null, false)'),
      'any type allowed (but not omitted)');

  // NodeList
  assert_equals(await nonNumericPolicy('document.getElementsByTagName(\"div\").item(x)'),
      'any type allowed (but not omitted)');
}, "Node*");

promise_test(async () => {
  // Range
  assert_equals(await nonNumericPolicy('document.createRange().setStart(document, x)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('document.createRange().setEnd(document, x)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('document.createRange().comparePoint(document, x)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('document.createRange().isPointInRange(document, x)'),
      'any type allowed (but not omitted)');
}, "Range");

promise_test(async () => {
  // Selection
  assert_equals(await nonNumericPolicy('getSelection().collapse(document, x)'),
      'any type allowed');
  assert_equals(await nonNumericPolicy('getSelection().setBaseAndExtent(document, x, document, 0)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('getSelection().setBaseAndExtent(document, 0, document, x)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('getSelection().collapse(document, x)'),
      'any type allowed');
  assert_equals(await nonNumericPolicy('getSelection().extend(document, x)'),
      'any type allowed');
  assert_equals(await nonNumericPolicy('getSelection().getRangeAt(x)'),
      'any type allowed (but not omitted)');
}, "Selection");

promise_test(async () => {
  // StyleSheetList
  assert_equals(await nonNumericPolicy('document.styleSheets.item(x)'),
      'any type allowed (but not omitted)');
}, "StyleSheetList");

promise_test(async () => {
  // Text
  assert_equals(await nonNumericPolicy('document.createTextNode(\"a\").splitText(x)'),
      'any type allowed (but not omitted)');
}, "Text");

promise_test(async () => {
  // TreeWalker
  assert_equals(await nonNumericPolicy('document.createTreeWalker(document, x, null, false)'),
      'any type allowed (but not omitted)');
}, "TreeWalker");

promise_test(async () => {
  // UIEvent
  assert_equals(await nonNumericPolicy('document.createEvent(\"UIEvent\").initUIEvent(\"a\", false, false, null, x)'),
      'any type allowed');
}, "UIEvent");

promise_test(async () => {
  // Window
  assert_equals(await nonNumericPolicy('window.scrollBy(x, 0)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('window.scrollBy(0, x)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('window.scrollTo(x, 0)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('window.scrollTo(0, x)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('window.scroll(x, 0)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('window.scroll(0, x)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('window.moveBy(x, 0)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('window.moveBy(0, x)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('window.moveTo(x, 0)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('window.moveTo(0, x)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('window.resizeBy(x, 0)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('window.resizeBy(0, x)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('window.resizeTo(x, 0)'),
      'any type allowed (but not omitted)');
  assert_equals(await nonNumericPolicy('window.resizeTo(0, x)'),
      'any type allowed (but not omitted)');
} ,"Window");

/*
This comment block is copied from the legacy test without checking the
availability of all APIs.

Non-tested APIs:
- History: go.
- TimeRanges: start, end.

Here are other examples of numeric types in function parameters and settable
attributes that we could test:
- CanvasGradient.idl: void addColorStop(in float offset, in DOMString color);
- XPathResult.idl: Node snapshotItem(in unsigned long index)
- SVGAngle.idl: in float valueInSpecifiedUnits);
- SVGLength.idl: in float valueInSpecifiedUnits);
- SVGMatrix.idl: [Custom] SVGMatrix translate(in float x, in float y);
- SVGMatrix.idl: [Custom] SVGMatrix scale(in float scaleFactor);
- SVGMatrix.idl: [Custom] SVGMatrix scaleNonUniform(in float scaleFactorX, in float scaleFactorY);
- SVGMatrix.idl: [Custom] SVGMatrix rotate(in float angle);
- SVGMatrix.idl: [Custom] SVGMatrix rotateFromVector(in float x, in float y)
- SVGMatrix.idl: [Custom] SVGMatrix skewX(in float angle);
- SVGMatrix.idl: [Custom] SVGMatrix skewY(in float angle);
- SVGNumber.idl: interface [Conditional=SVG, PODType=float] SVGNumber {
- SVGTransform.idl: void setTranslate(in float tx, in float ty);
- SVGTransform.idl: void setScale(in float sx, in float sy);
- SVGTransform.idl: void setRotate(in float angle, in float cx, in float cy);
- SVGTransform.idl: void setSkewX(in float angle);
- SVGTransform.idl: void setSkewY(in float angle);
- SVGAngle.idl: attribute float          value;
- SVGAngle.idl: attribute float          valueInSpecifiedUnits;
- SVGLength.idl: attribute float          value;
- SVGLength.idl: attribute float          valueInSpecifiedUnits;
- SVGNumber.idl: attribute float value
- SVGPoint.idl: attribute float x
- SVGPoint.idl: attribute float y
- SVGRect.idl: attribute float x
- SVGRect.idl: attribute float y
- SVGRect.idl: attribute float width
- SVGRect.idl: attribute float height
- SVGMatrix.idl: attribute double a;
- SVGMatrix.idl: attribute double b;
- SVGMatrix.idl: attribute double c;
- SVGMatrix.idl: attribute double d;
- SVGMatrix.idl: attribute double e;
- SVGMatrix.idl: attribute double f;
*/
</script>
